fix(shared): move getFileCategory to utils for proper export
- Move getFileCategory and isPreviewable from types to utils
- Add utils/portfolio.ts for file category utilities
- Ensure function is properly exported from @shadowob/shared
